+++ app/src/main/luaLibs/autotheme.lua
@@ -1,4 +1,0 @@
-local function autotheme()
-  return R.style.AppTheme_Material3
-end
-return autotheme