+++ app/src/main/resources/lua/autotheme.lua
@@ -1,0 +1,4 @@
+local function autotheme()
+  return R.style.Theme_AndroLuaX_Material3
+end
+return autotheme